refactor(scheduler): keep runtime adapter internal
