+++ rlinf/scheduler/__init__.py
@@ -20,6 +20,7 @@ from .hardware import (
     AcceleratorUtil,
     DOSW1HWInfo,
     DualFrankaHWInfo,
+    EmbodiedRuntimeCLI,
     FrankaHWInfo,
     GimArmHWInfo,
     HardwareInfo,
@@ -54,6 +55,7 @@ __all__ = [
     "AcceleratorType",
     "HardwareInfo",
     "DOSW1HWInfo",
+    "EmbodiedRuntimeCLI",
     "DualFrankaHWInfo",
     "FrankaHWInfo",
     "GimArmHWInfo",
