+++ rlinf/scheduler/hardware/__init__.py
@@ -25,6 +25,7 @@ from .robots import (
     DOSW1HWInfo,
     DualFrankaConfig,
     DualFrankaHWInfo,
+    EmbodiedRuntimeCLI,
     FrankaConfig,
     FrankaHWInfo,
     GimArmConfig,
@@ -48,6 +49,7 @@ __all__ = [
     "DOSW1HWInfo",
     "DualFrankaConfig",
     "DualFrankaHWInfo",
+    "EmbodiedRuntimeCLI",
     "FrankaConfig",
     "FrankaHWInfo",
     "GimArmConfig",
